mv db* to db/; move wizard_* to wizard/
