+++ src/db/dbcheck.cpp
